==> notebooks/note/data_prepare/4_prepare_RL.ipynb <==
from pathlib import Path
import os
os.getcwd()

import sys

module_path = os.path.abspath(os.path.join(os.getcwd(), '..', '..'))
sys.path.insert(0, module_path)

import note.base as base
sys.path.pop(0)

SEED = 42
base.fix_seed(SEED)

import zipfile

zip_path = Path('../../data/RLdata/RIKEN_LIPIDOMICS_dataset.zip')
extract_to = Path('../../data/RLdata/RIKEN_LIPIDOMICS_dataset')
with zipfile.ZipFile(zip_path, 'r') as zip_ref:
    zip_ref.extractall(extract_to)

extract_to = Path('../../data/RLdata/RIKEN_LIPIDOMICS_dataset')
for path in list(extract_to.rglob('*.xlsx')):
    dataframe = base.loadDataframe_for_RL(str(path))
    base.insertdata(dataframe, path='../../data/RLdata/ms2_rikenlipid.db')

def get_fileid(path):
    path_elements = path.split("/")
    file_name = path_elements[-1]
    file_name_elements = file_name.split("_")
    return file_name_elements[0]

import pandas as pd

df = base.import_sqlite3_data('../../data/RLdata/ms2_rikenlipid.db').drop_duplicates().reset_index(drop=True)
df['dataset'] = df['dataset'].str.replace('../../data/RLdata/RIKEN_LIPIDOMICS_dataset/dataset/', '', regex=False)
df['id'] = df['dataset'].apply(get_fileid).astype('int')
df = df[~df['Ontology'].str.contains('IS \(')]
df['Comment'] = df['Comment'].fillna('NC')
df = df[~(df['Comment'].str.contains('IS \(') | df['Comment'].str.contains('d') | df['Comment'].str.contains('two peaks'))].reset_index(drop=True)

df_pos = df[df['Adducttype'].str.contains('\]\+')  |  df['Adducttype'].str.contains('\]2\+')]
df_neg = df[df['Adducttype'].str.contains('\]\-') |  df['Adducttype'].str.contains('\]2\-')]
print(len(df), len(df_pos)+len(df_neg), len(df_neg), len(df_pos))

df_neg2 = df_neg[['Metabolitename', 'Ontology', 'id', 'AlignmentID','AverageMz', 'MSMSspectrum']].rename(columns={'id':'dataset'}).dropna().drop_duplicates().reset_index(drop=True)
df_pos2 = df_pos[['Metabolitename', 'Ontology', 'id', 'AlignmentID','AverageMz', 'MSMSspectrum']].rename(columns={'id':'dataset'}).dropna().drop_duplicates().reset_index(drop=True)

df_neg_exp_ = base.spectrum_tidy(df_neg2)
df_neg_neuloss_ = base.neutralloss_tidy(df_neg2)
df_neg_data = base.spectrum_neutralloss_table(df_neg_exp_, df_neg_neuloss_)    
df_neg_data['Ontology'] = df_neg_data['Ontology'].str.replace(r'-', '_')

df_pos_exp_ = base.spectrum_tidy(df_pos2)
df_pos_neuloss_ = base.neutralloss_tidy(df_pos2)
df_pos_data = base.spectrum_neutralloss_table(df_pos_exp_, df_pos_neuloss_)    
df_pos_data['Ontology'] = df_pos_data['Ontology'].str.replace(r'-', '_')

df_pos_data.to_csv('../../data/RLdata/df_RLtable_pos.csv', index=False) 
df_neg_data.to_csv('../../data/RLdata/df_RLtable_neg.csv', index=False)
