==> notebooks/note/data_prepare/5_prepare_HC.ipynb <==
import sys
import os

module_path = os.path.abspath(os.path.join(os.getcwd(), '..', '..'))
sys.path.insert(0, module_path)

import note.base as base
sys.path.pop(0)

SEED = 42
base.fix_seed(SEED)

import pandas as pd

pos

df_pos_height = pd.read_csv('../../data/HCdata/pos_Height_1_2024_03_29_19_43_37.txt', sep='\t',header=4, delimiter=None)\
[['Alignment ID','Average Mz', 'Metabolite name', 'Ontology','MS/MS spectrum']].\
rename(columns={'Metabolite name':'Metabolitename', 'Alignment ID':'AlignmentID','Average Mz':'AverageMz', 'MS/MS spectrum':'MSMSspectrum'}).dropna(subset=['MSMSspectrum']).reset_index(drop=True)
df_pos_height['dataset'] = 'HumanCohort'

df_pos_exp_ = base.spectrum_tidy(df_pos_height)
df_pos_neuloss_ = base.neutralloss_tidy(df_pos_height)
df_pos_data = base.spectrum_neutralloss_table(df_pos_exp_, df_pos_neuloss_)

neg

df_neg_height = pd.read_csv('../../data/HCdata/neg_Height_1_2024_03_29_19_47_56.txt', sep='\t',header=4, delimiter=None)\
[['Alignment ID','Average Mz', 'Metabolite name', 'Ontology','MS/MS spectrum']].\
rename(columns={'Metabolite name':'Metabolitename', 'Alignment ID':'AlignmentID','Average Mz':'AverageMz', 'MS/MS spectrum':'MSMSspectrum'}).dropna(subset=['MSMSspectrum']).reset_index(drop=True)
df_neg_height['dataset'] = 'HumanCohort'

df_neg_exp_ = base.spectrum_tidy(df_neg_height)
df_neg_neuloss_ = base.neutralloss_tidy(df_neg_height)
df_neg_data = base.spectrum_neutralloss_table(df_neg_exp_, df_neg_neuloss_)

df_pos_data.to_csv('../../data/HCdata/df_HCtable_pos.csv', index=False) 
df_neg_data.to_csv('../../data/HCdata/df_HCtable_neg.csv', index=False)
